+++ tests/PhpWord/Tests/Writer/ODTextTest.php
@@ -79,7 +79,7 @@ class ODTextTest extends \PHPUnit_Framework_TestCase
         $section->addLink('http://test.com');
         $section->addTitle('Test', 1);
         $section->addPageBreak();
-        $section->addTable();
+        $section->addTable()->addRow()->addCell()->addText('Test');
         $section->addListItem('Test');
         $section->addImage($imageSrc);
         $section->addObject($objectSrc);