+++ CHANGELOG.md
@@ -34,6 +34,7 @@ This release marked the change of PHPWord license from LGPL 2.1 to LGPL 3. Four 
 - Element: New `Line` element - @basjan GH-253
 - Title: Ability to apply numbering in heading - @ivanlanin GH-193
 - HTML Reader: Basic HTML reader - @ivanlanin GH-80 GH-254
+- RTF Writer: Basic table writing - @ivanlanin GH-245
 
 ### Bugfixes